indexer/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.2
easydict==1.13
flatbuffers==25.9.23
fonttools==4.61.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
httplib2==0.31.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.3.1
onnx==1.20.0
onnxruntime==1.16.3
opencv-python==4.7.0.72
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
prettytable==3.17.0
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rawpy==0.25.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.3
six==1.17.0
stringzilla==4.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.14
